+++ servers/media-server/src/server/gateway.rs
@@ -3,23 +3,24 @@ use std::{sync::Arc, time::Duration};
 use async_std::stream::StreamExt;
 use clap::Parser;
 use cluster::{
+    atm0s_sdn::{Publisher, PubsubSdk},
     implement::NodeId,
     rpc::{
-        gateway::{NodePing, NodePong, QueryBestNodesResponse},
+        gateway::{NodePing, QueryBestNodesResponse},
         general::{MediaEndpointCloseRequest, MediaEndpointCloseResponse, MediaSessionProtocol, NodeInfo, ServerType},
         webrtc::{WebrtcPatchRequest, WebrtcPatchResponse, WebrtcRemoteIceRequest, WebrtcRemoteIceResponse},
-        RpcEmitter, RpcEndpoint, RpcRequest, RPC_MEDIA_ENDPOINT_CLOSE, RPC_NODE_PING, RPC_WEBRTC_CONNECT, RPC_WEBRTC_ICE, RPC_WEBRTC_PATCH, RPC_WHEP_CONNECT, RPC_WHIP_CONNECT,
+        RpcEmitter, RpcEndpoint, RpcRequest, RPC_MEDIA_ENDPOINT_CLOSE, RPC_WEBRTC_CONNECT, RPC_WEBRTC_ICE, RPC_WEBRTC_PATCH, RPC_WHEP_CONNECT, RPC_WHIP_CONNECT,
     },
-    Cluster, ClusterEndpoint, GLOBAL_GATEWAY_SERVICE, INNER_GATEWAY_SERVICE, MEDIA_SERVER_SERVICE,
+    Cluster, ClusterEndpoint, MEDIA_SERVER_SERVICE,
 };
 use futures::{select, FutureExt};
-use media_utils::{SystemTimer, Timer, F32};
+use media_utils::{hash_str, SystemTimer, Timer, F32};
 use metrics::describe_counter;
 use metrics_dashboard::{build_dashboard_route, DashboardOptions};
 use poem::{web::Json, Route};
 use poem_openapi::OpenApiService;
 
-use crate::rpc::http::HttpRpcServer;
+use crate::{rpc::http::HttpRpcServer, server::gateway::logic::RouteResult};
 
 #[cfg(feature = "embed-samples")]
 use crate::rpc::http::EmbeddedFilesEndpoint;
@@ -39,7 +40,6 @@ use self::{
     rpc::{cluster::GatewayClusterRpc, http::GatewayHttpApis, RpcEvent},
 };
 
-pub use self::logic::GatewayMode;
 use super::MediaServerContext;
 
 mod ip2location;
@@ -54,14 +54,6 @@ const GATEWAY_SESSIONS_CONNECT_ERROR: &str = "gateway.sessions.connect.error";
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
 pub struct GatewayArgs {
-    /// Gateway mode
-    #[arg(value_enum, env, long, default_value_t = GatewayMode::Inner)]
-    pub mode: GatewayMode,
-
-    /// Gateway group, only set if mode is Inner
-    #[arg(env, long, default_value = "")]
-    pub group: String,
-
     /// lat location
     #[arg(env, long, default_value_t = 0.0)]
     pub lat: f32,
@@ -75,7 +67,16 @@ pub struct GatewayArgs {
     pub geoip_db: String,
 }
 
-pub async fn run_gateway_server<C, CR, RPC, REQ, EMITTER>(http_port: u16, http_tls: bool, opts: GatewayArgs, ctx: MediaServerContext<()>, cluster: C, rpc_endpoint: RPC) -> Result<(), &'static str>
+pub async fn run_gateway_server<C, CR, RPC, REQ, EMITTER>(
+    http_port: u16,
+    http_tls: bool,
+    zone: &str,
+    opts: GatewayArgs,
+    ctx: MediaServerContext<()>,
+    cluster: C,
+    rpc_endpoint: RPC,
+    pubsub: PubsubSdk,
+) -> Result<(), &'static str>
 where
     C: Cluster<CR> + Send + 'static,
     CR: ClusterEndpoint + Send + 'static,
@@ -86,10 +87,7 @@ where
     let node_id = cluster.node_id();
     let mut rpc_endpoint = GatewayClusterRpc::new(rpc_endpoint);
     let mut http_server: HttpRpcServer<RpcEvent> = crate::rpc::http::HttpRpcServer::new(http_port, http_tls);
-    let ip2location = match opts.mode {
-        GatewayMode::Global => Some(ip2location::Ip2Location::new(&opts.geoip_db)),
-        GatewayMode::Inner => None,
-    };
+    let ip2location = ip2location::Ip2Location::new(&opts.geoip_db);
 
     let timer = Arc::new(SystemTimer());
     let api_service = OpenApiService::new(GatewayHttpApis, "Gateway Server", env!("CARGO_PKG_VERSION")).server("/");
@@ -122,25 +120,49 @@ where
 
     http_server.start(route, ctx.clone()).await;
     let mut tick = async_std::stream::interval(Duration::from_millis(100));
-    let mut gateway_logic = GatewayLogic::new(opts.mode);
+    let mut gateway_logic = GatewayLogic::new(zone);
     let rpc_emitter = rpc_endpoint.emitter();
     let mut gateway_feedback_tick = async_std::stream::interval(Duration::from_millis(2000));
-    let dest_service_id = match opts.mode {
-        GatewayMode::Global => INNER_GATEWAY_SERVICE,
-        GatewayMode::Inner => MEDIA_SERVER_SERVICE,
-    };
+    let gateway_zone_channel_pub = pubsub.create_publisher(hash_str(&format!("gateway-zone-{}", zone)) as u32);
+    let gateway_zone_channel = pubsub.create_consumer(hash_str(&format!("gateway-zone-{}", zone)) as u32, None);
+    let gateway_channel_pub = pubsub.create_publisher(hash_str("gateway") as u32);
+    let gateway_channel = pubsub.create_consumer(hash_str("gateway") as u32, None);
 
     loop {
         let rpc = select! {
             _ = tick.next().fuse() => {
                 gateway_logic.on_tick(timer.now_ms());
                 continue;
-            }
-            _ = gateway_feedback_tick.next().fuse() => {
-                if matches!(opts.mode, GatewayMode::Inner) {
-                    ping_global_gateway(&gateway_logic, &opts.group, (F32::<2>::new(opts.lat), F32::<2>::new(opts.lng)), node_id, &rpc_emitter);
+            },
+            e = gateway_zone_channel.recv().fuse() => match e {
+                Some((_, from, _, data)) => {
+                    if from == node_id {
+                        continue;
+                    }
+                    if let Ok(ping) = bincode::deserialize(&data) {
+                        log::info!("[Gateway] node ping {:?}", ping);
+                        gateway_logic.on_node_ping(timer.now_ms(), &ping);
+                    }
+                    continue;
+                },
+                None => {
+                    continue;
                 }
-
+            },
+            e = gateway_channel.recv().fuse() => match e {
+                Some((_, _, _, data)) => {
+                    if let Ok(ping) = bincode::deserialize(&data) {
+                        log::info!("[Gateway] gateway ping {:?}", ping);
+                        gateway_logic.on_gateway_ping(timer.now_ms(), &ping);
+                    }
+                    continue;
+                },
+                None => {
+                    continue;
+                }
+            },
+            _ = gateway_feedback_tick.next().fuse() => {
+                ping_other_gateways(&gateway_logic, zone, (F32::<2>::new(opts.lat), F32::<2>::new(opts.lng)), node_id, &gateway_channel_pub);
                 continue;
             },
             rpc = http_server.recv().fuse() => {
@@ -154,12 +176,13 @@ where
         match rpc {
             RpcEvent::NodePing(req) => {
                 log::info!("[Gateway] node ping {:?}", req.param());
-                req.answer(Ok(gateway_logic.on_ping(timer.now_ms(), req.param())));
+                gateway_zone_channel_pub.send(bincode::serialize(req.param()).expect("Should serialize").into());
+                req.answer(Ok(gateway_logic.on_node_ping(timer.now_ms(), req.param())));
             }
-            RpcEvent::BestNodest(req) => {
+            RpcEvent::BestNodes(req) => {
                 log::info!("[Gateway] best nodes {:?}", req.param());
-                let nodes = gateway_logic.best_nodes(
-                    ip2location.as_ref().map(|f| f.get_location(&req.param().ip_addr)).flatten(),
+                let route_res = gateway_logic.best_nodes(
+                    ip2location.get_location(&req.param().ip_addr),
                     match req.param().protocol {
                         MediaSessionProtocol::Rtmp => ServiceType::Rtmp,
                         MediaSessionProtocol::Sip => ServiceType::Sip,
@@ -171,11 +194,15 @@ where
                     80,
                     req.param().size,
                 );
-                req.answer(Ok(QueryBestNodesResponse { nodes }));
+                if let RouteResult::OtherNode { nodes, service_id } = route_res {
+                    req.answer(Ok(QueryBestNodesResponse { nodes, service_id }));
+                } else {
+                    req.answer(Err("NOT_FOUND"));
+                }
             }
             RpcEvent::WhipConnect(req) => {
                 log::info!("[Gateway] whip connect compressed_sdp: {:?}", req.param().compressed_sdp.as_ref().map(|sdp| sdp.len()));
-                let location = ip2location.as_ref().map(|f| f.get_location(&req.param().ip_addr)).flatten();
+                let location = ip2location.get_location(&req.param().ip_addr);
                 webrtc_route::route_to_node(
                     rpc_emitter.clone(),
                     timer.clone(),
@@ -189,12 +216,11 @@ where
                     &req.param().user_agent.clone(),
                     req.param().session_uuid,
                     req,
-                    dest_service_id,
                 );
             }
             RpcEvent::WhepConnect(req) => {
                 log::info!("[Gateway] whep connect compressed_sdp: {:?}", req.param().compressed_sdp.as_ref().map(|sdp| sdp.len()));
-                let location = ip2location.as_ref().map(|f| f.get_location(&req.param().ip_addr)).flatten();
+                let location = ip2location.get_location(&req.param().ip_addr);
                 webrtc_route::route_to_node(
                     rpc_emitter.clone(),
                     timer.clone(),
@@ -208,12 +234,11 @@ where
                     &req.param().user_agent.clone(),
                     req.param().session_uuid,
                     req,
-                    dest_service_id,
                 );
             }
             RpcEvent::WebrtcConnect(req) => {
                 log::info!("[Gateway] webrtc connect compressed_sdp: {:?}", req.param().compressed_sdp.as_ref().map(|sdp| sdp.len()));
-                let location = ip2location.as_ref().map(|f| f.get_location(&req.param().ip_addr)).flatten();
+                let location = ip2location.get_location(&req.param().ip_addr);
                 webrtc_route::route_to_node(
                     rpc_emitter.clone(),
                     timer.clone(),
@@ -221,13 +246,12 @@ where
                     node_id,
                     ServiceType::Webrtc,
                     RPC_WEBRTC_CONNECT,
-                    req.param().ip_addr,
+                    req.param().ip_addr.clone().into(),
                     location,
                     &req.param().version.clone(),
                     &req.param().user_agent.clone(),
-                    req.param().session_uuid.expect("Should assign session_uuid on gateway"),
+                    req.param().session_uuid,
                     req,
-                    dest_service_id,
                 );
             }
             RpcEvent::WebrtcRemoteIce(req) => {
@@ -273,21 +297,16 @@ where
     }
 }
 
-fn ping_global_gateway<EMITTER: RpcEmitter + Send + 'static>(logic: &GatewayLogic, group: &str, location: (F32<2>, F32<2>), node_id: NodeId, rpc_emitter: &EMITTER) {
+fn ping_other_gateways(logic: &GatewayLogic, zone: &str, location: (F32<2>, F32<2>), node_id: NodeId, publisher: &Publisher) {
     let stats = logic.stats();
     let req = NodePing {
         node_id,
-        group: group.to_string(),
+        zone: zone.to_string(),
         location: Some(location),
         rtmp: stats.rtmp,
         sip: stats.sip,
         webrtc: stats.webrtc,
     };
 
-    let rpc_emitter = rpc_emitter.clone();
-    async_std::task::spawn(async move {
-        if let Err(e) = rpc_emitter.request::<_, NodePong>(GLOBAL_GATEWAY_SERVICE, None, RPC_NODE_PING, req, 1000).await {
-            log::error!("[Gateway] ping global gateway error {:?}", e);
-        }
-    });
+    publisher.send(bincode::serialize(&req).expect("Should serialize").into());
 }