+++ transports/sip/CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1](https://github.com/8xFF/atm0s-media-server/compare/atm0s-media-server-transport-sip-v0.1.0...atm0s-media-server-transport-sip-v0.1.1) - 2024-01-26
+
+### Other
+- updated the following local packages: atm0s-media-server-utils, atm0s-media-server-transport
+
 ## [0.1.0](https://github.com/8xFF/atm0s-media-server/releases/tag/atm0s-media-server-transport-sip-v0.1.0) - 2023-11-23
 
 ### Fixed